+++ Changelog.md
@@ -1,1 +1,17 @@
 # Changelog for WingetTools
+
+## v0.3.0
+
++ Added `Get-WGInstalled` to list winget installed packages.
++ Added `Source` parameter to `Get-WGInstalled` and `Get-WGPackage`.
++ Updated `README.md`.
++ Update about help topic.
+
+## v0.2.0
+
++ Added parameter to `Get-WGPackage` to accept winget output.
++ Created external help documentation.
+
+## v0.1.0
+
++ initial module.